requirements.txt
psycopg2
tenacity
python-dotenv